+++ main.js
@@ -26,13 +26,11 @@ const writeFile = (filepath, data) => {
   };
 
   const content = `
-    // case 1:
-    import square from './square';
-    console.log("Area of square: ", square(3, 5));
+    function test(user) {
+        console.log(user);
+    };
 
-    // case 2:
-    import {a, b} from './square';
-    console.log("Area of square: ", a(3, 5));
+    export default test;
   `;
   const ast = contentToAST(content);
   const { transformedAst, transformedContent } = transform(ast, content);
